Matrix: use size_t instead of int

+++ include/unicy/Matrix.hpp
@@ -7,20 +7,20 @@
 namespace Unicy
 {
 
-template <typename T, int R, int C>
+template <typename T, size_t R, size_t C>
 class Matrix
 {
 public:
     std::array<std::array<T, C>, R> values = {std::array<T, C>()};
 
-    template <typename Ty = T, int Ry = R, int Cy = C>
+    template <typename Ty = T, size_t Ry = R, size_t Cy = C>
     static constexpr Matrix<Ty, Ry, Cy> identity()
     {
         static_assert(Ry == Cy, "Identity matrix can only be generated for square matrix.");
 
         Matrix<Ty, Ry, Cy> mat;
 
-        for (int i = 0; i < R && i < C; i++)
+        for (size_t i = 0; i < R && i < C; i++)
         {
             mat[i][i] = 1;
         }
@@ -35,7 +35,7 @@ public:
         return os << r.to_string();
     }
 
-    template <typename Tl, int Rl, int Cl, typename Tr, int Rr, int Cr>
+    template <typename Tl, size_t Rl, size_t Cl, typename Tr, size_t Rr, size_t Cr>
     friend auto operator*(const Matrix<Tl, Rl, Cl> &l, const Matrix<Tr, Rr, Cr> &r) -> Matrix<decltype(std::declval<Tl &>() * std::declval<Tr &>()), Rl, Cr>
     {
         static_assert(Cl == Rr, "Right matrix rows count me be equal to left matrix columns count.");
@@ -44,13 +44,13 @@ public:
 
         Matrix<MulT, Rl, Cr> mat;
 
-        for (int i = 0; i < Rl; i++)
+        for (size_t i = 0; i < Rl; i++)
         {
-            for (int j = 0; j < Cr; j++)
+            for (size_t j = 0; j < Cr; j++)
             {
                 auto value = MulT();
 
-                for (int k = 0; k < Rr; k++)
+                for (size_t k = 0; k < Rr; k++)
                 {
                     value += l[i][k] * r[k][j];
                 }
@@ -62,14 +62,14 @@ public:
         return mat;
     }
 
-    template <typename Tl, int Rl, int Cl, typename Ry>
+    template <typename Tl, size_t Rl, size_t Cl, typename Ry>
     friend auto operator*(const Matrix<Tl, Rl, Cl> &l, const Ry &r) -> Matrix<decltype(std::declval<Tl &>() * std::declval<Ry &>()), Rl, Cl>
     {
         Matrix<decltype(std::declval<Tl &>() * std::declval<Ry &>()), Rl, Cl> mat;
 
-        for (int y = 0; y < Rl; y++)
+        for (size_t y = 0; y < Rl; y++)
         {
-            for (int x = 0; x < Cl; x++)
+            for (size_t x = 0; x < Cl; x++)
             {
                 mat[y][x] = l[y][x] * r;
             }
@@ -78,7 +78,7 @@ public:
         return mat;
     }
 
-    template <typename Tr, int Rr, int Cr>
+    template <typename Tr, size_t Rr, size_t Cr>
     Matrix<T, R, C> &operator*=(const Matrix<Tr, Rr, Cr> &r)
     {
         Matrix<T, R, C> mat = static_cast<Matrix<T, Cr, R>>(*this * r);
@@ -96,14 +96,14 @@ public:
         return *this;
     }
 
-    template <typename Ty, int Ry, int Cy>
+    template <typename Ty, size_t Ry, size_t Cy>
     operator Matrix<Ty, Ry, Cy>()
     {
         auto mat = Matrix<Ty, Ry, Cy>();
 
-        for (int y = 0; y < Ry; y++)
+        for (size_t y = 0; y < Ry; y++)
         {
-            for (int x = 0; x < Cy; x++)
+            for (size_t x = 0; x < Cy; x++)
             {
                 mat[y][x] = static_cast<Ty>((*this)[y][x]);
             }
@@ -112,12 +112,12 @@ public:
         return mat;
     }
 
-    inline std::array<T, C> &operator[](int i)
+    inline std::array<T, C> &operator[](size_t i)
     {
         return this->values[i];
     }
 
-    inline const std::array<T, C> &operator[](int i) const
+    inline const std::array<T, C> &operator[](size_t i) const
     {
         return this->values[i];
     }
